Adding capability up to 10 committee members
Different Cal Poly Masters programs require various numbers of committee members. For example, 2019 Aerospace requires 3 other than your chair. 
System used above is linked to the number of members in the frontmatter.tex file. This fix could be universal, but works for up to 10 members which seems like large enough that it will never need to be changed.